added new parameters

+++ funcs.sh
@@ -208,12 +208,20 @@ exit_if_hotfix () {
 deployment_cleanup () {
   local script_dir_name=$1
   local build_id=$2
+  local app_name=$3
+  local other_file=$4
 
   [ -z "$script_dir_name" ] && echo "provide the 'scripts' directory" && exit 1;
+  [ -z "$app_name" ] && echo "provide the '.bin/go-build' file" && exit 1;
+  [ -z "$other_file" ] && echo "See Dockerfile for other files included during docker-build" && exit 1;
 
   if [[ $build_id != "" ]] ;then
       for dir in $(ls); do
-          if [[ $dir != $script_dir_name ]]; then
+          if [[ $dir != $script_dir_name ]] &&
+             [[ $dir != $app_name ]] &&
+             [[ $dir != "Dockerfile" ]] &&
+             [[ $dir != $other_file ]];
+          then
               rm -rf $dir
           fi
       done
